Login UI Rework #1.2
