+++ spoti-vote/src/components/Login/LoginButton.jsx
@@ -7,6 +7,7 @@ const portFront = window.location.port || 80;
 const portBack = 8888;
 
 let defaultStyle = {
+	// color: constants.colors.font,
 	padding: "17px 48px",
 	fontSize: "0.9em",
 	lineHeight: 1,
@@ -22,12 +23,39 @@ let defaultStyle = {
 }
 class LoginButton extends Component {
 
+	constructor() {
+		super();
+		this.state = {
+			hover: false
+		}
+	}
+
+	toggleHover() {
+		this.setState({
+			hover: !this.state.hover
+		})
+	}
+
 	login() {
 		window.location.href = 'http://' + ipAddress + ':' + portBack + '/login';
 	}
 
 	render() {
-		return (<button id="loginbutton" style={defaultStyle} onClick={this.login.bind(this)} tabIndex="0">
+		let linkStyle;
+		if (this.state.hover) {
+			linkStyle = {
+				cursor: 'pointer',
+				backgroundColor: constants.colors.greenHover
+			}
+		} else {
+			linkStyle = {
+				cursor: 'context-menu'
+			}
+		}
+		return (<button id="loginbutton" style={{
+				...defaultStyle,
+				...linkStyle
+			}} onMouseEnter={this.toggleHover.bind(this)} onMouseLeave={this.toggleHover.bind(this)} onClick={this.login.bind(this)} tabIndex="0">
 			Host
 		</button>);
 	}
