+++ spoti-vote/src/index.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Login from './components/Login.jsx';
-import Login2 from './components/Login2.jsx';
+import Login_old from './components/Login_old.jsx';
+import Join from './components/Join.jsx';
 import App from './components/App.jsx';
 import './css/reset.css';
 import './css/fonts.css';
@@ -14,7 +15,8 @@ ReactDOM.render((<BrowserRouter>
 	<div>
 		<Route exact={true} path="/" component={Login}/>
 		<Route path="/app" component={App}/>
-		<Route path="/newLogin" component={Login2}/>
+		<Route path="/oldLogin" component={Login_old}/>
+		<Route path="/join" component={Join}/>
 	</div>
 </BrowserRouter>), document.getElementById('root'));
 registerServiceWorker();
